+++ libft/ft_strcmp_rev.c
@@ -6,13 +6,13 @@
 /*   By: bbarakov <bbarakov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/06 12:18:25 by bbarakov          #+#    #+#             */
-/*   Updated: 2014/12/10 13:27:09 by bbarakov         ###   ########.fr       */
+/*   Updated: 2014/12/11 19:58:25 by bbarakov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strcmp_rev(void *str1, void *str2)
+long		ft_strcmp_rev(void *str1, void *str2)
 {
 	int			i;
 	char		*s1;
